Move new enum to its own export (eg, for use in clients)
Move the rest of errors.ts to two separately named files.
