+++ packages/server/src/errors/index.ts
@@ -1,0 +1,9 @@
+export enum ApolloServerErrorCode {
+  GRAPHQL_PARSE_FAILED = 'GRAPHQL_PARSE_FAILED',
+  GRAPHQL_VALIDATION_FAILED = 'GRAPHQL_VALIDATION_FAILED',
+  PERSISTED_QUERY_NOT_FOUND = 'PERSISTED_QUERY_NOT_FOUND',
+  PERSISTED_QUERY_NOT_SUPPORTED = 'PERSISTED_QUERY_NOT_SUPPORTED',
+  BAD_USER_INPUT = 'BAD_USER_INPUT',
+  OPERATION_RESOLUTION_FAILURE = 'OPERATION_RESOLUTION_FAILURE',
+  BAD_REQUEST = 'BAD_REQUEST',
+}
