+++ packages/server/src/index.ts
@@ -1,5 +1,3 @@
 export { ApolloServer } from './ApolloServer.js';
 // Note that this is purely a type export.
 export * from './externalTypes/index.js';
-
-export { ApolloServerErrorCode } from './errors.js';
